requirements.txt
pyniryo2
opencv-python
pillow
keyboard